requirements.txt
# ----- ML & data tools ------
numpy==1.23.3
pandas==1.3
text_preprocessing==0.1.1
scikit-learn>=0.18
catboost==1.1.1
# ----- Micoservices ------
Flask==2.2.2
flask_restful==0.3.9
flask-autodoc==0.1.2
Flask==2.2.2
flask_restful==0.3.9
flask-autodoc==0.1.2
flask_restful_swagger==0.20.1
PyYAML==6.0
# ------ data labeling ------
label-studio==1.5.0
# --------------------------------
# visualisation
# --------------------------------
umap-learn==0.5.3
streamlit==1.13.0
altair==4.2.0
matplotlib==3.6.0
jupyter==1.0.0
